Move 'NewsletterContext' to '/react'
